[XHAB-11] implements splash screen navigation, refactors code
