+++ lib/src/data/userRepository.dart
@@ -1,0 +1,9 @@
+import 'package:xhabits/src/domain/splashScreenUseCase.dart';
+
+class UserRepository implements SplashScreenUseCase {
+  @override
+  bool isUserSignedIn() {
+    // TODO: implement isUserSignedIn https://trello.com/c/RF95DP0n/37-implement-isusersignedin
+    return true;
+  }
+}
