+++ lib/src/domain/splashScreenUseCase.dart
