+++ lib/src/presentation/scenes/home/home_screen.dart
@@ -1,0 +1,15 @@
+import 'package:flutter/material.dart';
+
+class HomeScreen extends StatefulWidget {
+  @override
+  _HomeScreenState createState() => _HomeScreenState();
+}
+
+class _HomeScreenState extends State<HomeScreen> {
+
+  @override
+  Widget build(BuildContext context) {
+    return Text('Home page');
+  }
+
+}
