--- lib/userRepository.dart
@@ -1,9 +1,0 @@
-import 'splashScreenUseCase.dart';
-
-class UserRepository implements SplashScreenUseCase {
-  @override
-  bool isUserSignedIn() {
-    // TODO: implement isUserSignedIn
-    return true;
-  }
-}
