+++ test/widget_test.dart
@@ -9,11 +9,12 @@ import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:xhabits/main.dart';
+import 'package:xhabits/src/presentation/XHApp.dart';
 
 void main() {
   testWidgets('Counter increments smoke test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
-    await tester.pumpWidget(MyApp());
+    await tester.pumpWidget(XHApp());
 
     // Verify that our counter starts at 0.
     expect(find.text('0'), findsOneWidget);
